addCoverage in bamCache.helper.js
